# docs/openAPI.yaml
openapi: 3.0.0
info:
  title: 家教系統 API
  version: 1.3.0
  description: >
    MVP 對齊需求 + 可擴展。完整時區支援，時間 ISO 8601（含時區）；查詢可帶 timezone 參數。
    - 時區：支援全球時區，資料庫存 UTC，API 自動轉換本地時間
    - 分頁：page=1、pageSize=20（<=100）
    - 排序：sort=field:asc|desc，多欄逗號分隔
    - 衝突 409，政策違反 422，權限 403
servers:
  - url: http://localhost:3001
    description: Development (Docker)
  - url: https://api.example.com
    description: Production
  - url: https://staging.api.example.com
    description: Staging
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  parameters:
    Page: { in: query, name: page, schema: { type: integer, minimum: 1, default: 1 } }
    PageSize: { in: query, name: pageSize, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
    Sort: { in: query, name: sort, schema: { type: string, description: "e.g. createdAt:desc,startAt:asc" } }
    From: { in: query, name: from, schema: { type: string, format: date-time } }
    To: { in: query, name: to, schema: { type: string, format: date-time } }
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                - BOOKING_CONFLICT
                - POLICY_VIOLATION
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - VALIDATION_ERROR
                - DUPLICATE
                - RATE_LIMITED
                - INSUFFICIENT_CREDITS
                - HOLD_EXPIRED
                - SETTLEMENT_BLOCKED
                - ALREADY_REVIEWED
                - CANCEL_NOT_ALLOWED
                - INTERNAL_ERROR
            message: { type: string }
            details:
              oneOf: [ { type: string }, { type: object }, { type: array } ]
      required: [error]
    Paginated:
      type: object
      properties: { page: {type: integer}, pageSize: {type: integer}, total: {type: integer} }
    Role: { type: string, enum: [admin, teacher, student] }
...
    MaterialItem:
      type: object
      properties:
        id: { type: string }
        type: { type: string, enum: [page, pdf] }
        title: { type: string }
        folderId: { type: string }
        content: { type: string, nullable: true }
        fileUrl: { type: string, format: uri, nullable: true }
        previewUrl: { type: string, format: uri, nullable: true }
        meta: { type: object, additionalProperties: true, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    BookingStatus:
      type: string
      enum: [scheduled, pending, pending_teacher, canceled, completed, noshow]
    BookingSource:
      type: string
      enum: [student, admin, teacher, system]
    BookingProgress:
      type: object
      properties:
        percent: { type: integer, minimum: 0, maximum: 100 }
        attendance: { type: string, enum: [unknown, attended, noshow], default: unknown }
        hasReport: { type: boolean }
        evidenceCount: { type: integer }
        settlementStatus: { type: string, enum: [pending, blocked, ready, settled], default: pending }
    BookingPostClass:
      type: object
      properties:
        teacherReportSubmittedAt: { type: string, format: date-time, nullable: true }
        teacherComment: { type: string, nullable: true }
        reportStatus: { type: string, enum: [none, submitted, verified], default: none }
        evidenceCount: { type: integer }
    BookingSummary:
      type: object
      properties:
        id: { type: string }
        teacher: { type: object, properties: { id: {type: string}, name: {type: string}, avatarUrl: {type: string, format: uri, nullable: true} } }
        student: { type: object, properties: { id: {type: string}, name: {type: string} } }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        status: { $ref: '#/components/schemas/BookingStatus' }
        material: { $ref: '#/components/schemas/MaterialItem' }
        lastMessageAt: { type: string, format: date-time, nullable: true }
        meetingUrl: { type: string, format: uri, nullable: true }
        source: { $ref: '#/components/schemas/BookingSource' }
    BookingDetail:
      allOf:
        - $ref: '#/components/schemas/BookingSummary'
        - type: object
          properties:
            courseTitle: { type: string, nullable: true }
            message: { type: string, nullable: true }
            messages:
              type: array
              items:
                type: object
                properties:
                  id: { type: string }
                  senderId: { type: string }
                  text: { type: string }
                  createdAt: { type: string, format: date-time }
            canReschedule: { type: boolean }
            canCancel: { type: boolean }
            cancelPolicy:
              type: object
              properties:
                freeBeforeHours: { type: integer, default: 24 }
                tiered:
                  type: array
                  items:
                    type: object
                    properties:
                      minHours: { type: integer }
                      maxHours: { type: integer }
                      cancelCardCost: { type: integer }
            settlement:
              type: object
              properties:
                status: { type: string, enum: [pending, blocked, ready, settled], default: pending }
                teacherUnitUSD: { type: number, nullable: true }
                payableAt: { type: string, format: date-time, nullable: true }
            postClass:
              $ref: '#/components/schemas/BookingPostClass'
            progress:
              $ref: '#/components/schemas/BookingProgress'
...
paths:
...
  /bookings:
    get:
      tags: [Bookings]
      summary: 我的預約清單
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: roleView
          schema: { type: string, enum: [student, teacher] }
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
        - in: query
          name: status
          schema: { type: string, enum: [upcoming, past, canceled, pending] }
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Paginated'
                  - type: object
                    properties: { items: { type: array, items: { $ref: '#/components/schemas/BookingSummary' } } }
    post:
      tags: [Bookings]
      summary: 建立預約
      description: >
        建立新的預約。startsAt 應為 ISO 8601 格式（含時區），例如: 2025-10-06T14:00:00+08:00
        系統會自動驗證教師可用性和時間衝突。
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [teacherId, startsAt]
              properties:
                teacherId: { type: string, description: "教師ID" }
                startsAt: { type: string, format: date-time, description: "預約開始時間 (ISO 8601格式)", example: "2025-10-06T14:00:00+08:00" }
                durationMinutes: { type: integer, default: 30, description: "預約時長（分鐘）" }
                timezone: { type: string, default: "Asia/Taipei", description: "時區", example: "Asia/Shanghai" }
                courseTitle: { type: string, nullable: true, description: "課程標題" }
                message: { type: string, nullable: true, description: "預約留言" }
                materialId: { type: string, nullable: true, description: "教材ID" }
                source: { type: string, default: "student", enum: [student, admin, teacher, system], description: "預約來源" }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/BookingDetail' } } } }
        '409': { description: 時間衝突, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '422': { description: 政策違反/無可用次卡, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /bookings/{id}:
    get:
      tags: [Bookings]
      summary: 取得預約詳情
      security: [{ bearerAuth: [] }]
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/BookingDetail' } } } }
        '404': { description: Not Found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /bookings/{id}/messages:
    post:
      tags: [Bookings]
      summary: 預約留言
      security: [{ bearerAuth: [] }]
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [text], properties: { text: { type: string } } }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { type: object, properties: { id: { type: string }, text: { type: string }, senderId: { type: string }, createdAt: { type: string, format: date-time } } }
  /bookings/{id}/reschedule:
    post:
      tags: [Bookings]
      summary: 改期
      security: [{ bearerAuth: [] }]
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [newStartsAt, durationMinutes]
              properties:
                newStartsAt: { type: string, format: date-time }
                durationMinutes: { type: integer, enum: [30] }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                  status: { type: string, enum: [scheduled, pending_teacher] }
                  startsAt: { type: string, format: date-time }
                  endsAt: { type: string, format: date-time }
        '409': { description: 衝突, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '422': { description: 政策違反, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /bookings/{id}/cancel:
    post:
      tags: [Cancellations]
      summary: 取消預約（依角色與距離上課時間套用取消政策）
      security: [{ bearerAuth: [] }]
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
                cause:
                  type: string
                  enum: [student_request, teacher_request, admin_force, technical_issue]
                waivePolicy: { type: boolean, default: false, description: 僅 admin 可用以免除政策/扣卡 }
      responses:
        '200':
          description: Canceled
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                  status: { type: string, enum: [canceled] }
                  refund:
                    type: object
                    nullable: true
                    properties:
                      lessonCardsReturned: { type: integer }
                      cancelCardsConsumed: { type: integer }
                      compensationGranted: { type: integer }
                      notes: { type: string, nullable: true }
        '422': { description: 政策違反或不可取消時間窗, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /bookings/{id}/ics:
    get:
      tags: [Bookings]
      summary: 下載 .ics
      security: [{ bearerAuth: [] }]
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses:
        '200':
          description: ICS file
          content:
            text/calendar:
              schema: { type: string }
  /bookings/{id}/confirm:
    post:
      tags: [Bookings]
      summary: 老師/管理確認預約
      security: [{ bearerAuth: [] }]
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses:
        '200':
          description: Confirmed
          content:
            application/json:
              schema: { type: object, properties: { id: { type: string }, status: { type: string, enum: [scheduled] } } }
...
  /bookings/{id}/evidence:
    get:
      tags: [Bookings]
      summary: 列出課後證據
      security: [{ bearerAuth: [] }]
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        url: { type: string, format: uri, nullable: true }
                        mimeType: { type: string }
                        size: { type: integer }
                        uploadedAt: { type: string, format: date-time }
                        uploadedBy:
                          type: object
                          properties:
                            id: { type: string }
                            role: { $ref: '#/components/schemas/Role' }
                            name: { type: string, nullable: true }
    post:
      tags: [Bookings]
      summary: 上傳課後證據並綁定預約（category=class_recording）
      security: [{ bearerAuth: [] }]
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                  url: { type: string, format: uri, nullable: true }
                  mimeType: { type: string }
                  size: { type: integer }
                  uploadedAt: { type: string, format: date-time }
  /bookings/{id}/evidence/{fileId}:
    delete:
      tags: [Bookings]
      summary: 移除課後證據
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: fileId, in: path, required: true, schema: { type: string } }
      responses:
        '204': { description: No Content }

      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rubrics]
              properties:
                rubrics:
                  type: object
                  properties:
                    assignment: { type: string, nullable: true }
                    vocabulary: { type: string, nullable: true }
                    grammar: { type: string, nullable: true }
                    pronunciation: { type: string, nullable: true }
                    coherence: { type: string, nullable: true }
                    performance: { type: string, nullable: true }
                evidenceUrls:
                  type: array
                  items: { type: string, format: uri }
      responses:
        '200':
          description: OK（回報完成才可進入結算）
          content:
            application/json:
              schema:
                type: object
                properties:
                  bookingId: { type: string }
                  settlementStatus: { type: string, enum: [ready] }
...
  /reports/admin:
    get:
      tags: [Reports]
      summary: Admin 報表：預約數、完成率、取消率、技術取消比例
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  bookings:
                    type: object
                    properties:
                      total: { type: integer }
                      completed: { type: integer }
                      canceled: { type: integer }
                      technicalCanceled: { type: integer }
                      completionRate: { type: number }
                      cancelRate: { type: number }
                      technicalCancelRate: { type: number }
  /reports/teacher:
    get:
      tags: [Reports]
      summary: Teacher 報表（僅自己）
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  bookings:
                    type: object
                    properties:
                      total: { type: integer }
                      completed: { type: integer }
                      canceled: { type: integer }
                      technicalCanceled: { type: integer }
                  earningsEstimateUSD: { type: number }
...